requirements.txt
# Discord bot core
discord.py==1.7.3
aiohttp<3.8
aiocache

# Data processing and analysis
numpy
pandas

# Plotting and visualization
matplotlib
seaborn
cycler
pycairo
PyNaCl

# Image processing
pillow

# Web scraping and parsing
lxml

# Environment variables
python-dotenv

# Web framework
fastapi
uvicorn

# Development and testing (optional)
pytest
